Entity Update

+++ autoparking/src/main/java/com/fsk/microservice/autoparking/repository/EmployeeService.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.fsk.microservice.autoparking.entity.Employee;
 
-public interface EmployeeService extends JpaRepository<Employee, Integer> {
+public interface EmployeeService extends JpaRepository<Employee, Long> {
 
 }
